Address #106 review: unify toolpath colormaps, float32 tables, turbo test, Apache-2.0 notice
- toolpath.py now colours beads through the exact 256-entry reference
  tables in _colormap_data (via TABLES), removing the coarse 11-stop
  approximation so bead colouring matches ViewerClient scalar colouring.
- _colormap_data.TABLES are now float32; client._apply_colormap keeps the
  interpolation single-precision, dropping a redundant float64 pass over
  large color arrays.
- test_colormap: pin turbo first/mid/last table entries.
- Add THIRD_PARTY_NOTICES.md (turbo colormap is Apache-2.0, Google 2019;
  viridis/plasma CC0) and ship it in the sdist + wheel; reference it from
  the _colormap_data docstring.

+++ src/threejs_viewer/client.py
@@ -2009,7 +2009,9 @@ class ViewerClient:
             normalized = np.zeros_like(values)
         else:
             normalized = (values - cmin) / (cmax - cmin)
-        normalized = np.clip(normalized, 0, 1)
+        # float32 keeps the interpolation math single-precision to match the
+        # float32 tables (the result is float32 anyway).
+        normalized = np.clip(normalized, 0, 1).astype(np.float32)
 
         cmap_arr = TABLES.get(colormap, TABLES["viridis"])
         n_colors = len(cmap_arr)
